added degrees and minutes to degrees converter function

+++ src/gnss/gnss.h
@@ -1,3 +1,11 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include <stdint.h>  // needed for int8_t and int16_t
+#include <ctype.h>
+#include <string.h>
+#include "pico/stdlib.h"
+#include "hardware/uart.h"
+#include "hardware/irq.h"
 
 void parse_buffer(char *buffer, char **sentences);
 int parse_line(char *string, char **fields, int num_fields);
@@ -8,6 +16,6 @@ int hex2int(char *c);
 int hexchar2int(char c);
 void parse_zda(char **zda_msg, int16_t *year, int8_t *month, int8_t *day,
                int8_t *hour, int8_t *min, int8_t *sec);
-void parse_gga(char **gga_msg, float *latitude, int *north,
-               float *longitude, int *east);
+void parse_gga(char **gga_msg, double *latitude, int *north,
+               double *longitude, int *east);
